restructured tests; replaced separate propensity function
