+++ vivarium_public_health/risks/__init__.py
@@ -1,5 +1,5 @@
 from .distributions import get_distribution
 from .effect import RiskEffect, RiskEffectSet
-from .base_risk import RiskComponent
+from .base_risk import Risk
